custom colors, removed admin address, singular extension, removed unnecessary
