+++ HMSAdmin/Admin/AdminProfileView.swift
@@ -13,7 +13,7 @@ struct AdminProfileView: View {
                     presentationMode.wrappedValue.dismiss()
                 }) {
                     Text("Done")
-                        .foregroundColor(Color(hex: "006666"))
+                        .foregroundColor(Color.customPrimary)
                         .padding()
                 }
             }
@@ -27,7 +27,7 @@ struct AdminProfileView: View {
             Spacer().frame(height: 60)
 
             Circle()
-                .fill(Color(hex: "006666"))
+                .fill(Color.customPrimary)
                 .frame(width: 100, height: 100)
                 .overlay(
                     Text(admin1.initials)
@@ -80,7 +80,7 @@ struct AdminProfileView: View {
 
             Spacer().frame(height: 20)
         }
-        .background(Color(.systemGray5))
+        .background(Color.customBackground)
     }
 }
