+++ HMSAdmin/Doctor/Task.swift
@@ -7,10 +7,4 @@
 
 import SwiftUI
 
-extension Date{
-    static func updateHour(_ value: Int) -> Date{
-        let calendar = Calendar.current
-        return calendar.date(byAdding: .hour, value: value, to: .init()) ?? .init()
-    }
-}
